HS-32 Improve chordStyle generating perfomance
